requirements.txt
# Python Libraries
Flask==1.1.2
Django==3.2.5
openai==0.27.0
requests==2.25.1
numpy==1.21.0
pandas==1.3.0

# Database Libraries
psycopg2-binary==2.9.1
pymongo==3.11.4

# Frontend Libraries
react==17.0.2
angular==12.1.1

# Other Libraries
gunicorn==20.1.0
python-dotenv==0.17.1
